+++ resources/Scripts/basic/time_scheme/improved_euler.cl
@@ -83,10 +83,10 @@ __kernel void predictor(__global int* imove,
                         __global vec* dudt_in,
                         __global float* rho_in,
                         __global float* drhodt_in,
-                        unsigned int N,
+                        usize N,
                         float dt)
 {
-    unsigned int i = get_global_id(0);
+    usize i = get_global_id(0);
     if(i >= N)
         return;
 
@@ -131,10 +131,10 @@ __kernel void corrector(__global int* imove,
                         __global float* drhodt,
                         __global vec* dudt_in,
                         __global float* drhodt_in,
-                        unsigned int N,
+                        usize N,
                         float dt)
 {
-    unsigned int i = get_global_id(0);
+    usize i = get_global_id(0);
     if(i >= N)
         return;
