+++ resources/Scripts/cfd/Boundary/GP/PreInteractions.cl
@@ -65,7 +65,7 @@
  */
 __kernel void entry(const __global uint* iset,
                     const __global int* imove,
-                    const __global uint* associations,
+                    const __global usize* associations,
                     const __global vec* r,
                     const __global vec* normal,
                     const __global vec* dudt,
@@ -76,17 +76,17 @@ __kernel void entry(const __global uint* iset,
                     __global float* p,
                     __global vec* u,
                     __constant float* visc_dyn,
-                    uint N,
+                    usize N,
                     vec g)
 {
-    const uint i = get_global_id(0);
+    const usize i = get_global_id(0);
     if(i >= N)
         return;
     if(imove[i] != -1)
         return;
 
     // Let's get the associated boundary element, and check it is valid
-    const uint iref = associations[i];
+    const usize iref = associations[i];
     if(iref >= N)
         return;
     const vec_xyz n_i = normal[iref].XYZ;
